# openapi/secpilot.openapi.yaml
openapi: 3.1.0
info:
  title: SecPilot Defense API
  version: 1.0.0
  description: Telegram-first defensive cybersecurity monitoring, audit, and defense API.
servers:
  - url: http://localhost:8000
paths:
  /health:
    get:
      summary: Liveness check alias
      responses:
        "200":
          description: OK
  /health/live:
    get:
      summary: Liveness check
      responses:
        "200":
          description: OK
  /ready:
    get:
      summary: Readiness check alias
      responses:
        "200":
          description: Ready
        "503":
          description: Not ready
  /health/ready:
    get:
      summary: Readiness check
      responses:
        "200":
          description: Ready
        "503":
          description: Not ready
  /api/v1/assets:
    get:
      summary: List authorized assets
      security:
        - apiKeyAuth: []
      parameters:
        - name: include_deleted
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Asset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
        "401":
          description: Unauthorized
    post:
      summary: Create authorized asset
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetCreate"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "401":
          description: Unauthorized
        "409":
          description: Duplicate asset
        "422":
          description: Invalid target
  /api/v1/assets/{asset_id}:
    get:
      summary: Get authorized asset
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Asset not found
    patch:
      summary: Update authorized asset
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Asset updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Asset not found
    delete:
      summary: Soft delete authorized asset
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "204":
          description: Asset soft deleted
        "404":
          description: Asset not found
  /api/v1/dns/audit:
    post:
      summary: Defensive DNS audit for an active whitelisted domain or URL asset
      description: Passive DNS record collection only. Requires the asset to be active and whitelisted.
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetActionRequest"
      responses:
        "200":
          description: DNS audit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DnsAuditResult"
        "401":
          description: Unauthorized
        "403":
          description: Scope denied
        "404":
          description: Asset not found
        "429":
          description: Rate limit exceeded
  /api/v1/ssl/audit:
    post:
      summary: Defensive SSL/TLS audit for an active whitelisted domain or URL asset
      description: HTTPS certificate, TLS version, expiry, and HSTS summary only.
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetActionRequest"
      responses:
        "200":
          description: SSL/TLS audit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SslAuditResult"
        "401":
          description: Unauthorized
        "403":
          description: Scope denied
        "404":
          description: Asset not found
        "429":
          description: Rate limit exceeded
  /api/v1/subdomains/passive:
    post:
      summary: Passive subdomain discovery for an active whitelisted domain or URL asset
      description: Uses passive certificate transparency sources only. No brute force or active probing.
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetActionRequest"
      responses:
        "200":
          description: Passive subdomain discovery result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubdomainDiscoveryResult"
        "401":
          description: Unauthorized
        "403":
          description: Scope denied
        "404":
          description: Asset not found
        "429":
          description: Rate limit exceeded
  /api/v1/monitoring/status:
    post:
      summary: Monitoring status for an active whitelisted domain or URL asset
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetActionRequest"
      responses:
        "200":
          description: Monitoring status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoringStatus"
        "401":
          description: Unauthorized
        "403":
          description: Scope denied
        "404":
          description: Asset not found
  /api/v1/dns/{asset_id}:
    get:
      summary: Backward-compatible defensive DNS audit endpoint
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: DNS audit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DnsAuditResult"
  /api/v1/ssl/{asset_id}:
    get:
      summary: Backward-compatible defensive SSL/TLS audit endpoint
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: SSL/TLS audit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SslAuditResult"
  /api/v1/subdomains/{asset_id}:
    get:
      summary: Backward-compatible passive subdomain discovery endpoint
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: Passive subdomain discovery result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubdomainDiscoveryResult"
  /api/v1/monitoring/{asset_id}:
    get:
      summary: Backward-compatible monitoring status endpoint
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: Monitoring status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoringStatus"
  /api/v1/monitoring/{asset_id}/enable:
    post:
      summary: Enable defensive monitoring for an active whitelisted domain or URL asset
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: Monitoring enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoringStatus"
        "401":
          description: Unauthorized
        "403":
          description: Scope denied
        "404":
          description: Asset not found
  /api/v1/monitoring/{asset_id}/disable:
    post:
      summary: Disable defensive monitoring for an active whitelisted domain or URL asset
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: Monitoring disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoringStatus"
        "401":
          description: Unauthorized
        "403":
          description: Scope denied
        "404":
          description: Asset not found
  /api/v1/audits/web:
    post:
      summary: Roadmap create web audit job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebAuditRequest"
      responses:
        "202":
          description: Job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
        "403":
          description: Scope denied
  /api/v1/jobs/{job_id}:
    get:
      summary: Roadmap get job status
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
  /api/v1/incidents/{incident_id}/actions/block-ip:
    post:
      summary: Roadmap approval-gated source IP block
      security:
        - bearerAuth: []
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Action accepted
components:
  parameters:
    AssetId:
      name: asset_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    AssetCreate:
      type: object
      required:
        - asset_type
        - value
      properties:
        asset_type:
          type: string
          enum: [domain, ip, cidr, url]
        value:
          type: string
          maxLength: 512
        label:
          type:
            - string
            - "null"
          maxLength: 255
    AssetUpdate:
      type: object
      properties:
        label:
          type:
            - string
            - "null"
          maxLength: 255
        status:
          type: string
          enum: [active, deleted]
    Asset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_type:
          type: string
        value:
          type: string
        normalized_value:
          type: string
        label:
          type:
            - string
            - "null"
        status:
          type: string
        created_at:
          type: string
          format: date-time
    AssetActionRequest:
      type: object
      required:
        - asset_id
      properties:
        asset_id:
          type: string
          format: uuid
    DnsAuditResult:
      type: object
      properties:
        domain:
          type: string
        a_records:
          type: array
          items:
            type: string
        aaaa_records:
          type: array
          items:
            type: string
        mx_records:
          type: array
          items:
            type: string
        ns_records:
          type: array
          items:
            type: string
        txt_records:
          type: array
          items:
            type: string
        cname_records:
          type: array
          items:
            type: string
        spf:
          type: boolean
        dmarc:
          type: boolean
        asn:
          type:
            - string
            - "null"
        country:
          type:
            - string
            - "null"
        provider:
          type:
            - string
            - "null"
    SslAuditResult:
      type: object
      properties:
        domain:
          type: string
        issuer:
          type:
            - string
            - "null"
        expires_at:
          type:
            - string
            - "null"
        days_remaining:
          type:
            - integer
            - "null"
        tls_version:
          type:
            - string
            - "null"
        https_available:
          type: boolean
        hsts:
          type: boolean
    SubdomainDiscoveryResult:
      type: object
      properties:
        domain:
          type: string
        subdomains:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              first_seen:
                type:
                  - string
                  - "null"
        sources:
          type: array
          items:
            type: string
    MonitoringStatus:
      type: object
      properties:
        asset_id:
          type: string
          format: uuid
        enabled:
          type: boolean
        last_check_at:
          type:
            - string
            - "null"
          format: date-time
        last_status:
          type:
            - string
            - "null"
        last_detail:
          type:
            - object
            - "null"
          additionalProperties: true
        scheduler:
          type:
            - object
            - "null"
          additionalProperties: true
    WebAuditRequest:
      type: object
      required:
        - asset_id
        - target
      properties:
        asset_id:
          type: string
        target:
          type: string
          format: uri
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
    JobAccepted:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
          example: queued